+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.3 (2024-01-29)
 
 - fix socket leak on failed handshake https://github.com/plausible/ch/pull/153